+++ Eter/EarthMageElderbranch.h
@@ -18,6 +18,6 @@ public:
 	std::string getName() const;
 	std::string getDescription() const;
 
-	bool playMage(Board& board, Color color, int8_t value, int8_t x, int8_t y);
+	bool playMage(Board& board, Color color, int16_t value, int16_t x, int16_t y);
 };
