+++ Eter/PowerLava.cpp
@@ -16,10 +16,10 @@ std::string PowerLava::getDescription() const
 	return m_description;
 }
 
-void PowerLava::playPower(Board& board, Player& player1, Player& player2, uint8_t nrLvl)//nrLvl - numarul de putere petru o carte
+void PowerLava::playPower(Board& board, Player& player1, Player& player2, int16_t nrLvl)//nrLvl - numarul de putere petru o carte
 {
-	for (int i = 0; i < board.getSize(); i++) 
-		for (int j = 0; j < board.getSize(); j++)
+	for (int16_t i = 0; i < board.getSize(); i++) 
+		for (int16_t j = 0; j < board.getSize(); j++)
 		{
 			if (board[{i, j}].back().getValue() == nrLvl) {
 				if (ColorToString(board[{i, j}].back().getColor()) == "Red") {
@@ -39,15 +39,15 @@ void PowerLava::playPower(Board& board, Player& player1, Player& player2, uint8_
 		}
 }
 
-bool PowerLava::checkPower(Board& board, int8_t nrLvl)
+bool PowerLava::checkPower(Board& board, int16_t nrLvl)
 {
-	int count = 0;
-	for (int i = 0; i < board.getSize(); i++) {
+	int16_t count = 0;
+	for (int16_t i = 0; i < board.getSize(); i++) {
 		if (count >= 2)
 		{
 			return true;
 		}
-		for (int j = 0; j < board.getSize(); j++)
+		for (int16_t j = 0; j < board.getSize(); j++)
 		{
 			if (board[{i, j}].back().getValue() == nrLvl) {
 				count++;
